{{cookiecutter.project}}/{{cookiecutter.project_slug}}/{{cookiecutter.project_slug}}_app/app_utility/schemas.py
from pydantic import BaseModel

class StudentBase(BaseModel):
    sname: str | None = None
    saddr:  str

class StudentCreate(StudentBase):
    sclass: str

class Student(StudentBase):
    sno: str

    class Config:
        orm_mode = True